+++ examples/login3.rb
@@ -3,7 +3,17 @@ require 'glimmer-dsl-libui'
 class Login
   include Glimmer
   
-  attr_accessor :username, :password, :logged_in
+  attr_accessor :username, :password
+  attr_reader :logged_in
+  
+  def logged_in=(value)
+    @logged_in = value
+    self.logged_out = !value # calling logged_out= method notifies logged_out observers
+  end
+  
+  def logged_out=(value)
+    self.logged_in = !value unless logged_in == !value
+  end
   
   def logged_out
     !logged_in
@@ -18,19 +28,19 @@ class Login
           entry {
             label 'Username:'
             text <=> [self, :username]
-            enabled <= [self, :logged_out, computed_by: :logged_in] # computed_by option ensures being notified of changes to logged_in
+            enabled <= [self, :logged_out]
           }
           
           password_entry {
             label 'Password:'
             text <=> [self, :password]
-            enabled <= [self, :logged_out, computed_by: :logged_in]
+            enabled <= [self, :logged_out]
           }
         }
         
         horizontal_box {
           button('Login') {
-            enabled <= [self, :logged_out, computed_by: :logged_in]
+            enabled <= [self, :logged_out]
             
             on_clicked do
               self.logged_in = true