+++ examples/login4.rb
@@ -3,7 +3,13 @@ require 'glimmer-dsl-libui'
 class Login
   include Glimmer
   
-  attr_accessor :username, :password, :logged_in
+  attr_accessor :username, :password
+  attr_reader :logged_in
+  
+  def logged_in=(value)
+    @logged_in = value
+    notify_observers(:logged_out) # manually notify observers of logged_out upon logged_in changes; this method comes automatically from enhancement as Glimmer::DataBinding::ObservableModel via data-binding
+  end
   
   def logged_out
     !logged_in
@@ -18,19 +24,19 @@ class Login
           entry {
             label 'Username:'
             text <=> [self, :username]
-            enabled <= [self, :logged_out, computed_by: :logged_in] # computed_by option ensures being notified of changes to logged_in
+            enabled <= [self, :logged_out]
           }
           
           password_entry {
             label 'Password:'
             text <=> [self, :password]
-            enabled <= [self, :logged_out, computed_by: :logged_in]
+            enabled <= [self, :logged_out]
           }
         }
         
         horizontal_box {
           button('Login') {
-            enabled <= [self, :logged_out, computed_by: :logged_in]
+            enabled <= [self, :logged_out]
             
             on_clicked do
               self.logged_in = true